Add constructor for pre-existing DynamicWatcher
This allows users that already have a DynamicWatcher and are using Query
Batches to supply their instance to the template resolver. Most details
of managing the watcher and the batch are delegated to the user: this
package will not automatically stop the query batch after the template
is resolved, for instance.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>